fix: address code review feedback
- Fix IPC fallback: telegram-mcp now uses same fallback as harness
  (join(HOME || "/tmp", ".leoclaw", "ipc") instead of "/tmp/leo-ipc")
- Add mode 0o700 to all mkdirSync calls for IPC and tasks directories
- Log warning at startup when LEO_ALLOWED_CHAT_IDS is unset
- Cap ask_user timeout to 600s via z.number().max(600)
- Sanitize callback_data newlines before prompt construction
- Read skip-permissions from config.json in compile-crons.ts (env still
  takes precedence, matching harness behavior)
- Export buildChildEnv from utils.ts; import in harness and tests to
  eliminate logic duplication
