+++ src/utils.ts
@@ -15,6 +15,22 @@ export function parseBooleanEnv(value: string | undefined): boolean | undefined 
   return undefined;
 }
 
+export const ENV_ALLOWLIST = [
+  "PATH", "HOME", "USER", "LOGNAME", "SHELL", "TMPDIR", "LANG", "LC_ALL",
+  "TERM", "XDG_CONFIG_HOME", "XDG_DATA_HOME", "XDG_CACHE_HOME",
+  "NODE_PATH", "NODE_OPTIONS",
+  "LEO_IPC_DIR", "LEO_WORKSPACE",
+  "AGENT_BROWSER_PROFILE",
+];
+
+export function buildChildEnv(extra?: Record<string, string>): Record<string, string> {
+  const env: Record<string, string> = {};
+  for (const key of ENV_ALLOWLIST) {
+    if (process.env[key]) env[key] = process.env[key]!;
+  }
+  return { ...env, ...extra };
+}
+
 export function parseAllowedUsersEnv(value: string | undefined): string[] | undefined {
   if (!value) return undefined;
   return value
